* develop: (56 commits)
  Leverage BRANCH_NAME env variable on jenkins builds
  Test implicit flow behaviour
  Remove commented out code
  Remove useless dump headers filter
  docker: get branch name from env variable
  Fix missing license
  link license-check to validate phase
  Fix sonar badges
  More security configuration refactoring
  actuator: disable most actuator endpoints by default
  Fixed IdpId resolution in EPTID SAML resolver
  docker: push branch-latest tag
  More work on fixing EPTID resolution
  Typo in variable name
  Security configuration refactory
  Added license headers
  EduPersonTargetedID resolver and tests
  first attempt
  Add cleanup
  Add license header
  ...

+++ iam-persistence/src/main/java/it/infn/mw/iam/persistence/repository/IamOAuthAccessTokenRepository.java
@@ -59,24 +59,24 @@ public interface IamOAuthAccessTokenRepository
 
   @Query("select count(t) from OAuth2AccessTokenEntity t "
       + "where (t.expiration is NULL or t.expiration > :timestamp)")
-  int countValidAccessTokens(@Param("timestamp") Date timestamp);
+  long countValidAccessTokens(@Param("timestamp") Date timestamp);
 
   @Query("select count(t) from OAuth2AccessTokenEntity t "
       + "where (t.expiration is NULL or t.expiration > :timestamp) "
       + "and (t.authenticationHolder.userAuth.name = :userId)")
-  int countValidAccessTokensForUser(@Param("userId") String userId,
+  long countValidAccessTokensForUser(@Param("userId") String userId,
       @Param("timestamp") Date timestamp);
 
   @Query("select count(t) from OAuth2AccessTokenEntity t "
       + "where (t.expiration is NULL or t.expiration > :timestamp) "
       + "and (t.authenticationHolder.clientId = :clientId)")
-  int countValidAccessTokensForClient(@Param("clientId") String clientId,
+  long countValidAccessTokensForClient(@Param("clientId") String clientId,
       @Param("timestamp") Date timestamp);
 
   @Query("select count(t) from OAuth2AccessTokenEntity t "
       + "where (t.expiration is NULL or t.expiration > :timestamp) "
       + "and (t.authenticationHolder.userAuth.name = :userId) "
       + "and (t.authenticationHolder.clientId = :clientId)")
-  int countValidAccessTokensForUserAndClient(@Param("userId") String userId,
+  long countValidAccessTokensForUserAndClient(@Param("userId") String userId,
       @Param("clientId") String clientId, @Param("timestamp") Date timestamp);
 }
